chore: polishing + typo fixes
